Do it the Chris way
