--- lib/docker/omni_builder.rb
@@ -1,39 +1,0 @@
-module Docker
-  class OmniBuilder
-    delegate_missing_to :@runner
-    include Docker::Helpers
-    include Support::TempFileBlock
-    def initialize(runner)
-      @runner = runner
-    end
-
-    def run
-      temp_file_block(EntryFile(source_file), submission_dir) do |entry|
-        docker_run \
-          language_executable,
-          filename(entry),
-          chdir: submission_dir,
-          in: input_buffer
-      end
-    end
-
-    def language_executable
-      raise "#{self} - #{__method__} not implemented"
-    end
-
-    def build
-      MockResult.new
-    end
-
-    class MockResult
-      def success?; true end
-      def failure?; false end
-      def out; '' end
-      def err; '' end
-    end
-
-    private def filename(io_file)
-      io_file.path.rpartition('/').last
-    end
-  end
-end
