+++ spec/fixtures/submission_runners/general/ProblemA.in
