+++ spec/fixtures/submission_runners/general/bad/ProblemA.rb
