+++ spec/fixtures/submission_runners/general/corrupt_code.rb
