+++ spec/fixtures/submission_runners/general/good/ProblemA.rb
