+++ spec/lib/submission_runners/ruby_builder_spec.rb
@@ -1,10 +1,10 @@
 require 'rails_helper'
-require_relative '../../lib/docker/helpers'
+require_relative '../../../lib/docker/helpers'
 include Docker::Helpers
 
 RSpec.describe 'OmniBuilder with Ruby', type: 'docker' do
   describe 'docker command and barebone ruby image' do
-    let(:dir) { Dir.pwd + '/spec/fixtures/docker/' }
+    let(:dir) { Dir.pwd + '/spec/fixtures/submission_runners/general/' }
 
     let(:submission) {
       ->i, o, e {
